+++ c_src/godot/godot.spec.exs
@@ -6,7 +6,7 @@ state_type "MyState"
 
 spec init(state, arguments :: [string]) :: {:ok :: label, state, code :: int}
                                   | {:fail :: label, state, result :: string}
-spec iteration(state, delta :: int) :: {:ok :: label, state, code :: int}
+spec iteration(state, delta :: float) :: {:ok :: label, state, code :: int}
                                   | {:fail :: label, state, result :: string}
 spec call(state, method :: string)  :: {:ok :: label, state, :string :: label, result_string :: string}
                                   # | {:ok :: label, state, :float :: label, result_string :: float}
